file structure organized into folders and Router implemented
