+++ src/Components/App/App.css
@@ -15,7 +15,7 @@ svg{
 }
 .App{
   background-image: url("https://images.unsplash.com/photo-1495195129352-aeb325a55b65?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1510&q=80");
-  background-size: 1fr;
+  background-size: 100%;
   background-repeat: no-repeat;
 }
