+++ src/Components/Donut/Donut.js
@@ -1,5 +1,4 @@
 import React, { Component } from "react";
-import "./App.css";
 
 export class Donut extends Component {
   constructor() {
