+++ src/Components/NoMatch/NoMatch.js
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export default function NoMatch() {
+  return (
+    <section className="NoMatch">
+      <h3>Sorry! That page doesn't exist.</h3>
+      <button>Back</button>
+    </section>
+  )
+}
