+++ src/Components/Projects/Projects.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Projects() {
+  return (
+    <section>
+      
+    </section>
+  )
+}
